[#264] Adopt minio code distributed with AGPLv3 license in gateway components
